removed old, urllib2-based transport class, and renamed urllib3 transport
