+++ elasticapm/conf/__init__.py
@@ -143,7 +143,7 @@ class Config(_ConfigBase):
     hostname = _ConfigValue('HOSTNAME', default=socket.gethostname())
     auto_log_stacks = _BoolConfigValue('AUTO_LOG_STACKS', default=True)
     keyword_max_length = _ConfigValue('KEYWORD_MAX_LENGTH', type=int, default=1024)
-    transport_class = _ConfigValue('TRANSPORT_CLASS', default='elasticapm.transport.http_urllib3.AsyncUrllib3Transport',
+    transport_class = _ConfigValue('TRANSPORT_CLASS', default='elasticapm.transport.http.AsyncTransport',
                                    required=True)
     processors = _ListConfigValue('PROCESSORS', default=[
         'elasticapm.processors.sanitize_stacktrace_locals',
