+++ elasticapm/conf/defaults.py
@@ -79,8 +79,4 @@ ASYNC_MODE = True
 # Should elasticapm wrap middleware for better metrics detection
 INSTRUMENT_DJANGO_MIDDLEWARE = True
 
-SYNC_TRANSPORT_CLASS = 'elasticapm.transport.http_urllib3.Urllib3Transport'
-
-ASYNC_TRANSPORT_CLASS = 'elasticapm.transport.http_urllib3.AsyncUrllib3Transport'
-
 TIMESTAMP_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
