+++ elasticapm/transport/__init__.py
@@ -1,6 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from elasticapm.transport.http import AsyncHTTPTransport, HTTPTransport
-
-
-default = [HTTPTransport, AsyncHTTPTransport]
