+++ tests/transports/test_urllib3.py
@@ -5,7 +5,7 @@ from urllib3.exceptions import MaxRetryError, TimeoutError
 from urllib3_mock import Responses
 
 from elasticapm.transport.base import TransportException
-from elasticapm.transport.http_urllib3 import Urllib3Transport
+from elasticapm.transport.http import Transport
 from elasticapm.utils import compat
 
 try:
@@ -19,14 +19,14 @@ responses = Responses('urllib3')
 
 def test_send(httpserver):
     httpserver.serve_content(code=202, content='', headers={'Location': 'http://example.com/foo'})
-    transport = Urllib3Transport(urlparse.urlparse(httpserver.url))
+    transport = Transport(urlparse.urlparse(httpserver.url))
     url = transport.send(compat.b('x'), {})
     assert url == 'http://example.com/foo'
 
 
 @responses.activate
 def test_timeout():
-    transport = Urllib3Transport(urlparse.urlparse('http://localhost'))
+    transport = Transport(urlparse.urlparse('http://localhost'))
     responses.add('POST', '/', status=202,
                   body=MaxRetryError(None, None, reason=TimeoutError()))
     with pytest.raises(TransportException) as exc_info:
@@ -36,7 +36,7 @@ def test_timeout():
 
 def test_http_error(httpserver):
     httpserver.serve_content(code=418, content="I'm a teapot")
-    transport = Urllib3Transport(urlparse.urlparse(httpserver.url))
+    transport = Transport(urlparse.urlparse(httpserver.url))
 
     with pytest.raises(TransportException) as exc_info:
         transport.send('x', {})
@@ -49,7 +49,7 @@ def test_generic_error():
     url, status, message, body = (
         'http://localhost:9999', 418, "I'm a teapot", 'Nothing'
     )
-    transport = Urllib3Transport(urlparse.urlparse(url))
+    transport = Transport(urlparse.urlparse(url))
     responses.add('POST', '/', status=status, body=Exception('Oopsie'))
     with pytest.raises(TransportException) as exc_info:
         transport.send('x', {})
@@ -58,20 +58,20 @@ def test_generic_error():
 
 def test_http_proxy_environment_variable():
     with mock.patch.dict('os.environ', {'HTTP_PROXY': 'http://example.com'}):
-        transport = Urllib3Transport(urlparse.urlparse('http://localhost:9999'))
+        transport = Transport(urlparse.urlparse('http://localhost:9999'))
         assert isinstance(transport.http, urllib3.ProxyManager)
 
 
 def test_https_proxy_environment_variable():
     with mock.patch.dict('os.environ', {'HTTPS_PROXY': 'https://example.com'}):
-        transport = Urllib3Transport(urlparse.urlparse('http://localhost:9999'))
+        transport = Transport(urlparse.urlparse('http://localhost:9999'))
         assert isinstance(transport.http, urllib3.poolmanager.ProxyManager)
 
 
 def test_https_proxy_environment_variable_is_preferred():
     with mock.patch.dict('os.environ', {'HTTPS_PROXY': 'https://example.com',
                                         'HTTP_PROXY': 'http://example.com'}):
-        transport = Urllib3Transport(urlparse.urlparse('http://localhost:9999'))
+        transport = Transport(urlparse.urlparse('http://localhost:9999'))
         assert isinstance(transport.http, urllib3.poolmanager.ProxyManager)
         assert transport.http.proxy.scheme == 'https'
 
@@ -85,9 +85,9 @@ def test_header_encodings():
     headers = {
         compat.text_type('X'): compat.text_type('V')
     }
-    transport = Urllib3Transport(urlparse.urlparse('http://localhost:9999'))
+    transport = Transport(urlparse.urlparse('http://localhost:9999'))
 
-    with mock.patch('elasticapm.transport.http_urllib3.urllib3.PoolManager.urlopen') as mock_urlopen:
+    with mock.patch('elasticapm.transport.http.urllib3.PoolManager.urlopen') as mock_urlopen:
         mock_urlopen.return_value = mock.Mock(status=202)
         transport.send('', headers)
     _, args, kwargs = mock_urlopen.mock_calls[0]
